pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tview"
dynamic = ["version"]
description = "Publication-quality alignment viewer for nucleotide and amino acid sequences"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Troy Sincomb", email = "troysincomb@gmail.com"},
]
keywords = ["bioinformatics", "alignment", "samtools", "tview", "visualization", "BAM", "FASTA"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "matplotlib>=3.5",
    "click>=8.0",
    "pyyaml>=6.0",
    "pysam>=0.20; sys_platform != 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "hypothesis>=6.0",
]

[project.urls]
Homepage = "https://github.com/MurrellGroup/tview"
Repository = "https://github.com/MurrellGroup/tview"
Issues = "https://github.com/MurrellGroup/tview/issues"

[project.scripts]
tview = "tview.cli:main"

[tool.setuptools.package-data]
tview = ["*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_file = "src/tview/_version.py"
local_scheme = "no-local-version"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
known_third_party = ["matplotlib", "click", "pysam"]

[tool.pytest.ini_options]
testpaths = ["tests"]
